requirements.txt
asgiref==3.8.1
Django==5.2
django-cors-headers==4.7.0
django-debug-toolbar==5.1.0
django-extensions==3.2.3
django-filter==25.1
django-silk==5.4.0
django-tenants==3.8.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
environs==14.1.1
gprof2dot==2025.4.14
inflection==0.5.1
marshmallow==3.26.1
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
uritemplate==4.1.1
